[RENAME/#68] list->itmes 명칭 변경
