+++ app/src/main/java/com/napzak/market/data/explore/dto/ProductBuyItemsRequest.kt
@@ -4,13 +4,11 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ProductSellListRequest(
+data class ProductBuyItemsRequest(
     @SerialName("sortOption")
     val sortOption: String,
     @SerialName("genreId")
     val genreIds: List<Long>?,
     @SerialName("isOnSale")
     val isOnSale: Boolean,
-    @SerialName("isUnopened")
-    val isUnopened: Boolean,
 )
