+++ app/src/main/java/com/napzak/market/data/explore/dto/ProductSellItemsRequest.kt
@@ -4,11 +4,13 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ProductBuyListRequest(
+data class ProductSellItemsRequest(
     @SerialName("sortOption")
     val sortOption: String,
     @SerialName("genreId")
     val genreIds: List<Long>?,
     @SerialName("isOnSale")
     val isOnSale: Boolean,
+    @SerialName("isUnopened")
+    val isUnopened: Boolean,
 )
