+++ app/src/main/java/com/napzak/market/data/explore/mapper/ExploreProductBuyItemsMapper.kt
@@ -1,20 +1,19 @@
 package com.napzak.market.data.explore.mapper
 
-import com.napzak.market.data.explore.dto.ProductSellListRequest
-import com.napzak.market.data.explore.dto.ProductSellListResponse
+import com.napzak.market.data.explore.dto.ProductBuyItemsRequest
+import com.napzak.market.data.explore.dto.ProductBuyItemsResponse
 import com.napzak.market.domain.explore.model.ProductItem
 import com.napzak.market.domain.explore.model.ProductListFilter
 
-fun ProductListFilter.toProductSellRequest(): ProductSellListRequest =
-    ProductSellListRequest(
+fun ProductListFilter.toProductBuyRequest(): ProductBuyItemsRequest =
+    ProductBuyItemsRequest(
         sortOption = this.sortOption,
         isOnSale = this.isOnSale,
-        isUnopened = this.isUnopened,
         genreIds = this.genreId,
     )
 
-fun ProductSellListResponse.toProductSellList(): List<ProductItem> =
-    productSellList.map { productItem ->
+fun ProductBuyItemsResponse.toProductBuyList(): List<ProductItem> =
+    productBuyList.map { productItem ->
         with(productItem) {
             ProductItem(
                 productId = productId,
@@ -26,6 +25,7 @@ fun ProductSellListResponse.toProductSellList(): List<ProductItem> =
                 isInterested = isInterested,
                 tradeType = tradeType,
                 tradeStatus = tradeStatus,
+                isPriceNegotiable = isPriceNegotiable,
                 isOwnedByCurrentUser = isOwnedByCurrentUser,
             )
         }
